+++ scalus-core/jvm/src/test/scala/scalus/prelude/ShowTest.scala
@@ -54,7 +54,7 @@ class ShowTest extends AnyFunSuite with EvalTestKit {
         val uplc = sir.toUplc(optimizeUplc = true)
         val result = uplc.evaluateDebug
         assert(result.success.term == "{0: \"0011\"}".asTerm)
-        assert(result.budget == ExUnits(memory = 56369, steps = 13_324373))
+        assert(result.budget == ExUnits(memory = 56369, steps = 13326839))
         assertEvalEq(
           Rational(1, 2).toData.show,
           "<0, [1, 2]>" // Rational is represented as a pair of BigInts