+++ scalus-core/shared/src/test/scala/scalus/prelude/MathTest.scala
@@ -138,14 +138,14 @@ class MathTest extends AnyFunSuite with EvalTestKit:
           gcd(BigInt(12), BigInt(18)),
           BigInt(6),
           Seq(
-            compilerOptions -> ExUnits(memory = 15713, steps = 3525304)
+            compilerOptions -> ExUnits(memory = 15713, steps = 3526537)
           )
         )
         assertEvalWithBudgets(
           gcd(BigInt(-12), BigInt(18)),
           BigInt(6),
           Seq(
-            compilerOptions -> ExUnits(memory = 15713, steps = 3525304)
+            compilerOptions -> ExUnits(memory = 15713, steps = 3526537)
           )
         )
 
@@ -174,7 +174,7 @@ class MathTest extends AnyFunSuite with EvalTestKit:
           BigInt(17).sqRoot,
           BigInt(4),
           Seq(
-            compilerOptions -> ExUnits(memory = 35665, steps = 12811018)
+            compilerOptions -> ExUnits(memory = 35665, steps = 12814717)
           )
         )
         assertEvalFailsWithMessage[RequirementError]("sqrt: negative radicand")(BigInt(-1).sqRoot)
@@ -185,21 +185,21 @@ class MathTest extends AnyFunSuite with EvalTestKit:
           sqrt(BigInt("1000000000000")), // 10^12
           BigInt("1000000"),
           Seq(
-            compilerOptions -> ExUnits(memory = 37969, steps = 13421696)
+            compilerOptions -> ExUnits(memory = 37969, steps = 13425395)
           )
         )
         assertEvalWithBudgets(
           sqrt(BigInt("1000000000000000000000000000000")), // 10^30
           BigInt("1000000000000000"),
           Seq(
-            compilerOptions -> ExUnits(memory = 42676, steps = 14820672)
+            compilerOptions -> ExUnits(memory = 42676, steps = 14828070)
           )
         )
         assertEvalWithBudgets(
           sqrt(BigInt("1000000000000000000000000000000000000000000000000000000000000")), // 10^60
           BigInt("1000000000000000000000000000000"),
           Seq(
-            compilerOptions -> ExUnits(memory = 55213, steps = 18370167)
+            compilerOptions -> ExUnits(memory = 55213, steps = 18400170)
           )
         )
 
@@ -222,35 +222,35 @@ class MathTest extends AnyFunSuite with EvalTestKit:
           pow(BigInt(2), BigInt(3)),
           BigInt(8),
           Seq(
-            compilerOptions -> ExUnits(memory = 18928, steps = 4865297)
+            compilerOptions -> ExUnits(memory = 18928, steps = 4866941)
           )
         )
         assertEvalWithBudgets(
           pow(BigInt(-2), BigInt(3)),
           BigInt(-8),
           Seq(
-            compilerOptions -> ExUnits(memory = 18928, steps = 4865297)
+            compilerOptions -> ExUnits(memory = 18928, steps = 4866941)
           )
         )
         assertEvalWithBudgets(
           pow(BigInt(7), BigInt(2)),
           BigInt(49),
           Seq(
-            compilerOptions -> ExUnits(memory = 18124, steps = 4545136)
+            compilerOptions -> ExUnits(memory = 18124, steps = 4546780)
           )
         )
         assertEvalWithBudgets(
           pow(BigInt(513), BigInt(3)),
           BigInt(135005697),
           Seq(
-            compilerOptions -> ExUnits(memory = 18928, steps = 4865297)
+            compilerOptions -> ExUnits(memory = 18928, steps = 4866941)
           )
         )
         assertEvalWithBudgets(
           pow(BigInt(2), BigInt(42)),
           BigInt("4398046511104"),
           Seq(
-            compilerOptions -> ExUnits(memory = 42164, steps = 11215766)
+            compilerOptions -> ExUnits(memory = 42164, steps = 11220698)
           )
         )
 
@@ -266,35 +266,35 @@ class MathTest extends AnyFunSuite with EvalTestKit:
           exp2(BigInt(0)),
           BigInt(1),
           Seq(
-            compilerOptions -> ExUnits(memory = 5309, steps = 3758628)
+            compilerOptions -> ExUnits(memory = 5309, steps = 3759450)
           )
         )
         assertEvalWithBudgets(
           exp2(BigInt(1)),
           BigInt(2),
           Seq(
-            compilerOptions -> ExUnits(memory = 5309, steps = 3758628)
+            compilerOptions -> ExUnits(memory = 5309, steps = 3759450)
           )
         )
         assertEvalWithBudgets(
           exp2(BigInt(4)),
           BigInt(16),
           Seq(
-            compilerOptions -> ExUnits(memory = 5309, steps = 3758628)
+            compilerOptions -> ExUnits(memory = 5309, steps = 3759450)
           )
         )
         assertEvalWithBudgets(
           exp2(BigInt(42)),
           BigInt("4398046511104"),
           Seq(
-            compilerOptions -> ExUnits(memory = 5309, steps = 3758628)
+            compilerOptions -> ExUnits(memory = 5309, steps = 3759450)
           )
         )
         assertEvalWithBudgets(
           exp2(BigInt(256)),
           BigInt("115792089237316195423570985008687907853269984665640564039457584007913129639936"),
           Seq(
-            compilerOptions -> ExUnits(memory = 5319, steps = 3939663)
+            compilerOptions -> ExUnits(memory = 5319, steps = 3940485)
           )
         )
 
@@ -345,21 +345,21 @@ class MathTest extends AnyFunSuite with EvalTestKit:
           log(BigInt(10), base = BigInt(2)),
           BigInt(3),
           Seq(
-            compilerOptions -> ExUnits(memory = 16619, steps = 3944756)
+            compilerOptions -> ExUnits(memory = 16619, steps = 3945989)
           )
         )
         assertEvalWithBudgets(
           log(BigInt(42), base = BigInt(2)),
           BigInt(5),
           Seq(
-            compilerOptions -> ExUnits(memory = 23829, steps = 5805710)
+            compilerOptions -> ExUnits(memory = 23829, steps = 5807765)
           )
         )
         assertEvalWithBudgets(
           log(BigInt(42), base = BigInt(3)),
           BigInt(3),
           Seq(
-            compilerOptions -> ExUnits(memory = 16619, steps = 3944756)
+            compilerOptions -> ExUnits(memory = 16619, steps = 3945989)
           )
         )
         assertEvalWithBudgets(
@@ -373,7 +373,7 @@ class MathTest extends AnyFunSuite with EvalTestKit:
           log(BigInt(4), base = BigInt(4)),
           BigInt(1),
           Seq(
-            compilerOptions -> ExUnits(memory = 9409, steps = 2083802)
+            compilerOptions -> ExUnits(memory = 9409, steps = 2084213)
           )
         )
         assertEvalWithBudgets(