+++ scalus-examples/jvm/src/test/scala/scalus/examples/editablenft/EditableNftValidatorTest.scala
@@ -409,7 +409,7 @@ class EditableNftValidatorTest extends AnyFunSuite, ScalusTest {
           signer = Alice.signer
         )
         assertResult(
-          ExUnits(memory = 177417L, steps = 51827990L)
+          ExUnits(memory = 177417L, steps = 51824198L)
         ):
             burnTx.witnessSet.redeemers.get.value.totalExUnits
         val burnResult = provider.submit(burnTx).await()