+++ scalus-examples/jvm/src/test/scala/scalus/examples/escrow/EscrowTest.scala
@@ -109,7 +109,7 @@ class EscrowTest extends AnyFunSuite, ScalusTest {
         )
 
         assertResult(
-          ExUnits(memory = 231352L, steps = 79337104L)
+          ExUnits(memory = 231352L, steps = 79334280L)
         ):
             depositTx.witnessSet.redeemers.get.value.totalExUnits
         val result = provider.submit(depositTx).await()
@@ -214,7 +214,7 @@ class EscrowTest extends AnyFunSuite, ScalusTest {
         )
 
         assertResult(
-          ExUnits(memory = 211276L, steps = 70850794L)
+          ExUnits(memory = 211276L, steps = 70847970L)
         ):
             payTx.witnessSet.redeemers.get.value.totalExUnits
         val result = provider.submit(payTx).await()
@@ -299,7 +299,7 @@ class EscrowTest extends AnyFunSuite, ScalusTest {
         )
 
         assertResult(
-          ExUnits(memory = 228352L, steps = 79151493L)
+          ExUnits(memory = 228352L, steps = 79148669L)
         ):
             refundTx.witnessSet.redeemers.get.value.totalExUnits
         val result = provider.submit(refundTx).await()