+++ scalus-examples/jvm/src/test/scala/scalus/examples/lottery/LotteryValidatorTest.scala
@@ -46,7 +46,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           revealTx,
           lotteryUtxo._1,
-          ExUnits(memory = 139593L, steps = 47201240L)
+          ExUnits(memory = 139593L, steps = 47198043L)
         )
     }
 
@@ -73,7 +73,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           revealTx,
           lotteryUtxo._1,
-          ExUnits(memory = 140194L, steps = 47328113L)
+          ExUnits(memory = 140194L, steps = 47324916L)
         )
     }
 
@@ -100,7 +100,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           revealTx,
           lotteryUtxo._1,
-          ExUnits(memory = 139593L, steps = 47201240L)
+          ExUnits(memory = 139593L, steps = 47198043L)
         )
     }
 
@@ -241,7 +241,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           p2RevealTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 98733L, steps = 31826993L)
+          ExUnits(memory = 98733L, steps = 31825397L)
         )
     }
 
@@ -288,7 +288,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           timeoutTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 134687L, steps = 42188704L)
+          ExUnits(memory = 134687L, steps = 42185285L)
         )
     }
 
@@ -376,7 +376,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           loseTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 128310L, steps = 40616871L)
+          ExUnits(memory = 128310L, steps = 40613452L)
         )
     }
 
@@ -421,7 +421,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           timeoutTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 134687L, steps = 42233880L)
+          ExUnits(memory = 134687L, steps = 42230461L)
         )
     }
 
@@ -571,7 +571,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           loseTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 128310L, steps = 40662047L)
+          ExUnits(memory = 128310L, steps = 40658628L)
         )
     }
 
@@ -741,7 +741,7 @@ class LotteryValidatorTest extends AnyFunSuite, ScalusTest {
           provider,
           p1RevealTx,
           lotteryUtxo2._1,
-          ExUnits(memory = 98132L, steps = 31700120L)
+          ExUnits(memory = 98132L, steps = 31698524L)
         )
     }
 }